+++ src/arguments/gameSetting.ts
@@ -3,9 +3,13 @@ import type { GameSettings } from '@mafia/structures/Game';
 import { Argument, ArgumentResult } from '@sapphire/framework';
 
 export default class extends Argument<keyof GameSettings> {
-	public run(argument: string): ArgumentResult<keyof GameSettings> {
-		if (!Reflect.has(DEFAULT_GAME_SETTINGS, argument))
-			return this.error('ArgumentGameSettingInvalidKey', `"${argument}" is not a valid game setting.`);
-		return this.ok(argument as keyof GameSettings);
+	public run(parameter: string): ArgumentResult<keyof GameSettings> {
+		if (!Reflect.has(DEFAULT_GAME_SETTINGS, parameter))
+			return this.error({
+				identifier: 'ArgumentGameSettingInvalidKey',
+				message: `"${parameter}" is not a valid game setting.`,
+				parameter
+			});
+		return this.ok(parameter as keyof GameSettings);
 	}
 }