+++ src/arguments/piece.ts
@@ -1,11 +1,15 @@
 import { Argument, ArgumentResult, Piece } from '@sapphire/framework';
 
 export default class extends Argument<Piece> {
-	public run(argument: string): ArgumentResult<Piece> {
+	public run(parameter: string): ArgumentResult<Piece> {
 		for (const store of this.context.client.stores.values()) {
-			if (store.has(argument)) return this.ok(store.get(argument)!);
+			if (store.has(parameter)) return this.ok(store.get(parameter)!);
 		}
 
-		return this.error(argument, 'ArgumentPieceInvalidPiece', 'Invalid piece');
+		return this.error({
+			identifier: 'ArgumentPieceInvalidPiece',
+			message: 'Invalid piece',
+			parameter
+		});
 	}
 }