+++ src/arguments/player.ts
@@ -4,18 +4,27 @@ import { Argument, ArgumentContext, ArgumentStore, AsyncArgumentResult, Result, 
 import type { User } from 'discord.js';
 
 export default class extends Argument<Player> {
-	public async run(argument: string, context: PlayerArgumentContext): AsyncArgumentResult<Player> {
+	public async run(parameter: string, context: PlayerArgumentContext): AsyncArgumentResult<Player> {
 		const game = context.game ?? context.message.channel.game;
 
-		const resolved = Player.resolve(game!, argument);
+		const resolved = Player.resolve(game!, parameter);
 		if (resolved) return this.ok(resolved);
 
-		const user = (await ((this.store as unknown) as ArgumentStore).get('user')!.run(argument, context)) as Result<User, UserError>;
-		if (!user.success) return this.error(argument, 'ArgumentPlayerUnknownUser', 'Invalid player mentioned.');
+		const user = (await ((this.store as unknown) as ArgumentStore).get('user')!.run(parameter, context)) as Result<User, UserError>;
+		if (!user.success)
+			return this.error({
+				identifier: 'ArgumentPlayerUnknownUser',
+				message: 'Invalid player mentioned.',
+				parameter
+			});
 		const player = game!.players.get(user.value);
 
 		if (player) return this.ok(player);
-		return this.error(argument, 'ArgumentPlayerInvalidPlayer', 'That user is not playing.');
+		return this.error({
+			identifier: 'ArgumentPlayerInvalidPlayer',
+			message: 'That user is not playing.',
+			parameter
+		});
 	}
 }