+++ src/preconditions/DayOnly.ts
@@ -6,6 +6,8 @@ export default class extends Precondition {
 	public run(message: Message) {
 		const { game } = message.channel;
 		if (game && game.phase === Phase.Day) return this.ok();
-		return this.error('This command cannot be used at night or during trials.');
+		return this.error({
+			message: 'This command cannot be used at night or during trials.'
+		});
 	}
 }